+++ madgui/line_view.py
@@ -2,16 +2,23 @@
 View component for the MadGUI application.
 """
 
+# GUI
+import wx
+
 # scipy
 import numpy as np
 import matplotlib as mpl
 from matplotlib.figure import Figure
 from matplotlib.ticker import MultipleLocator
 
+# 3rd party
 from obsub import event
+from cern.resource.package import PackageResource
 
+# internal
 from .model import Vector
 from .unit import units, stripunit, unit_label
+from .element_view import MadElementPopup, MadElementView
 
 
 class MadLineView(object):
@@ -21,6 +28,12 @@ class MadLineView(object):
     This is automatically updated when the model changes.
 
     """
+
+    @classmethod
+    def create(cls, model, frame):
+        view = cls(model)
+        frame.AddView(view, model.name)
+        return view
 
     def __init__(self, model):
         """Create a matplotlib figure and register as observer."""
@@ -196,6 +209,23 @@ class MirkoView(object):
     is replotted.
 
     """
+    ON_MIRKO = wx.NewId()
+
+    @classmethod
+    def connect_toolbar(cls, panel):
+        view = panel.view
+        mirko = cls(view.model, view)
+        bmp = wx.ArtProvider.GetBitmap(wx.ART_GO_HOME, wx.ART_TOOLBAR)
+        panel.toolbar.AddCheckTool(
+                cls.ON_MIRKO,
+                bitmap=bmp,
+                shortHelp='Show MIRKO envelope',
+                longHelp='Show MIRKO envelope for comparison. The envelope is computed for the default parameters.')
+        wx.EVT_TOOL(panel, cls.ON_MIRKO, mirko.OnMirkoClick)
+
+    def OnMirkoClick(self, event):
+        """Invoked when user clicks Mirko-Button"""
+        self.visible = event.IsChecked()
 
     def __init__(self, model, view):
         """
@@ -258,4 +288,139 @@ class MirkoView(object):
                 l.remove()
             self.lines = None
             self.view.figure.canvas.draw()
+
+
+class MadCtrl(object):
+    """
+    Controller class for a ViewPanel and MadModel
+
+    """
+    ON_MATCH = wx.NewId()
+    ON_SELECT = wx.NewId()
+
+    @classmethod
+    def create(cls, viewpanel):
+        return cls(viewpanel.view.model, viewpanel)
+
+
+    def __init__(self, model, panel):
+        """Initialize and subscribe as observer for user events."""
+        self.cid_match = None
+        self.cid_select = None
+        self.model = model
+        self.panel = panel
+        self.view = panel.view
+
+        # match
+        res = PackageResource(__package__)
+        with res.open(['resource', 'cursor.xpm']) as xpm:
+            img = wx.ImageFromStream(xpm, wx.BITMAP_TYPE_XPM)
+        bmp = wx.BitmapFromImage(img)
+        panel.toolbar.AddCheckTool(
+                self.ON_MATCH,
+                bitmap=bmp,
+                shortHelp='Beam matching',
+                longHelp='Match by specifying constraints for envelope x(s), y(s).')
+        wx.EVT_TOOL(panel, self.ON_MATCH, self.OnMatchClick)
+
+        # select
+        bmp = wx.ArtProvider.GetBitmap(wx.ART_TIP, wx.ART_TOOLBAR)
+        panel.toolbar.AddCheckTool(
+                self.ON_SELECT,
+                bitmap=bmp,
+                shortHelp='Show info for individual elements',
+                longHelp='Show info for individual elements')
+        wx.EVT_TOOL(panel, self.ON_SELECT, self.OnSelectClick)
+
+
+    def OnSelectClick(self, event):
+        """Invoked when user clicks Mirko-Button"""
+        if event.IsChecked():
+            self.start_select()
+        else:
+            self.stop_select()
+
+    def start_select(self):
+        """Start select mode."""
+        self.cid_select = self.view.figure.canvas.mpl_connect(
+                'button_press_event',
+                self.on_select)
+
+    def stop_select(self):
+        """Stop select mode."""
+        if self.cid_select is not None:
+            self.view.figure.canvas.mpl_disconnect(self.cid_select)
+            self.cid_select = None
+
+    def OnMatchClick(self, event):
+        """Invoked when user clicks Match-Button"""
+        if event.IsChecked():
+            self.start_match()
+        else:
+            self.stop_match()
+
+    def start_match(self):
+        """Start matching mode."""
+        self.cid_match = self.view.figure.canvas.mpl_connect(
+                'button_press_event',
+                self.on_match)
+        self.constraints = []
+
+    def stop_match(self):
+        """Stop matching mode."""
+        if self.cid_match is not None:
+            self.view.figure.canvas.mpl_disconnect(self.cid_match)
+            self.cid_match = None
+        self.model.clear_constraints()
+
+    @property
+    def frame(self):
+        wnd = self.panel
+        while wnd.GetParent():
+            wnd = wnd.GetParent()
+        return wnd
+
+    def on_select(self, event):
+        """Display a popup window with info about the selected element."""
+        elem = self.model.element_by_position_center(
+            event.xdata * self.view.unit.x)
+        if elem is None or 'name' not in elem:
+            return
+        popup = MadElementPopup(self.frame)
+        element_view = MadElementView(popup, self.model, elem['name'])
+        popup.Show()
+
+
+    def on_match(self, event):
+        axes = event.inaxes
+        if axes is None:
+            return
+        axis = 0 if axes is self.view.axes.x else 1
+
+        elem = self.model.element_by_position_center(
+            event.xdata * self.view.unit.x)
+        if elem is None or 'name' not in elem:
+            return
+
+        if event.button == 2:
+            self.model.remove_constraint(elem)
+            return
+        elif event.button != 1:
+            return
+
+        orig_cursor = self.panel.GetCursor()
+        wait_cursor = wx.StockCursor(wx.CURSOR_WAIT)
+        self.panel.SetCursor(wait_cursor)
+
+        # add the clicked constraint
+        envelope = event.ydata*self.view.unit.y
+        self.model.add_constraint(axis, elem, envelope)
+
+        # add another constraint to hold the orthogonal axis constant
+        orth_axis = 1-axis
+        orth_env = self.model.get_envelope_center(elem, orth_axis)
+        self.model.add_constraint(orth_axis, elem, orth_env)
+
+        self.model.match()
+        self.panel.SetCursor(orig_cursor)
