+++ madgui/mainframe.py
@@ -1,25 +1,12 @@
-# standard library
-import os
-
-# wxwidgets
+# GUI components
 import wx
 import wx.aui
 from wx.py.crust import Crust
-
-# matplotlib
 from matplotlib.backends.backend_wxagg import FigureCanvasWxAgg as Canvas
 from matplotlib.backends.backend_wxagg import NavigationToolbar2WxAgg as Toolbar
-from matplotlib.backends.backend_wx import _load_bitmap
-
-# 3rd party
-from obsub import event
-from cern import cpymad
-from cern.resource.package import PackageResource
 
 # project internal
-from .model import MadModel, Vector
-from .line_view import MadLineView, MirkoView
-from .controller import MadCtrl
+from .plugin import hookcollection
 
 
 #----------------------------------------
@@ -27,16 +14,24 @@ from .controller import MadCtrl
 #----------------------------------------
 
 class ViewPanel(wx.Panel):
+
     """
-    Display panel view for a MadLineView figure.
+    Display panel for a matplotlib figure.
     """
-    ON_MATCH = wx.NewId()
-    ON_MIRKO = wx.NewId()
-    ON_OPEN = wx.NewId()
-    ON_SELECT = wx.NewId()
+
+    hook = hookcollection(
+        'madgui.viewpanel', [
+            'init'
+        ])
 
     def __init__(self, parent, view, **kwargs):
-        """Initialize panel and connect the view."""
+
+        """
+        Initialize panel and connect the view.
+
+        Extends wx.App.__init__.
+        """
+
         super(ViewPanel, self).__init__(parent, **kwargs)
 
         self.view = view
@@ -45,45 +40,9 @@ class ViewPanel(wx.Panel):
         self.canvas = Canvas(self, -1, view.figure)
         view.canvas = self.canvas
 
+        # create a toolbar
         self.toolbar = Toolbar(self.canvas)
-
-        res = PackageResource(__package__)
-
-        with res.open(['resource', 'cursor.xpm']) as xpm:
-            img = wx.ImageFromStream(xpm, wx.BITMAP_TYPE_XPM)
-        bmp = wx.BitmapFromImage(img)
-        self.toolbar.AddCheckTool(
-                self.ON_MATCH,
-                bitmap=bmp,
-                shortHelp='Beam matching',
-                longHelp='Match by specifying constraints for envelope x(s), y(s).')
-        wx.EVT_TOOL(self, self.ON_MATCH, self.OnMatchClick)
-
-        bmp = wx.ArtProvider.GetBitmap(wx.ART_TIP, wx.ART_TOOLBAR)
-        self.toolbar.AddCheckTool(
-                self.ON_SELECT,
-                bitmap=bmp,
-                shortHelp='Show info for individual elements',
-                longHelp='Show info for individual elements')
-        wx.EVT_TOOL(self, self.ON_SELECT, self.OnSelectClick)
-
-        bmp = wx.ArtProvider.GetBitmap(wx.ART_GO_HOME, wx.ART_TOOLBAR)
-        self.toolbar.AddCheckTool(
-                self.ON_MIRKO,
-                bitmap=bmp,
-                shortHelp='Show MIRKO envelope',
-                longHelp='Show MIRKO envelope for comparison. The envelope is computed for the default parameters.')
-        wx.EVT_TOOL(self, self.ON_MIRKO, self.OnMirkoClick)
-
-        # TODO: this should not be within the notebook page:
-        bmp = wx.ArtProvider.GetBitmap(wx.ART_FILE_OPEN, wx.ART_TOOLBAR)
-        self.toolbar.AddSimpleTool(
-            self.ON_OPEN,
-            bitmap=bmp,
-            shortHelpString='Open another model',
-            longHelpString='Open another model')
-        wx.EVT_TOOL(self, self.ON_OPEN, self.OnOpenClick)
-
+        self.hook.init(self)
         self.toolbar.Realize()
 
         # put element into sizer
@@ -92,38 +51,23 @@ class ViewPanel(wx.Panel):
         sizer.Add(self.toolbar, 0 , wx.LEFT | wx.EXPAND)
         self.SetSizer(sizer)
 
-    @event
-    def OnMatchClick(self, event):
-        """Invoked when user clicks Match-Button"""
-        pass
-
-    @event
-    def OnMirkoClick(self, event):
-        """Invoked when user clicks Mirko-Button"""
-        pass
-
-    @event
-    def OnSelectClick(self, event):
-        """Invoked when user clicks Mirko-Button"""
-        pass
-
-    @event
-    def OnOpenClick(self, event):
-        """Invoked when user clicks Open-Model Button."""
-        wx.GetApp().open_model(self.GetParent())
-
     def OnPaint(self, event):
         """Handle redraw by painting canvas."""
         self.canvas.draw()
 
 
 class Frame(wx.Frame):
+
     """
     Main window.
     """
 
-    ID_SHELL = wx.NewId()
-    ID_MODEL = wx.NewId()
+    hook = hookcollection(
+        'madgui.frame', [
+            'init',
+            'term',
+            'menu'
+        ])
 
     @classmethod
     def create(cls, app):
@@ -132,19 +76,22 @@ class Frame(wx.Frame):
         return frame
 
     def __init__(self, logfolder):
-        """Create notebook frame."""
+
+        """
+        Create notebook frame.
+
+        Extends wx.Frame.__init__.
+        """
+
         super(Frame, self).__init__(
             parent=None,
             title='MadGUI',
-            size=wx.Size(800,600))
+            size=wx.Size(800, 600))
 
         self.logfolder = logfolder
 
         # create and listen to events:
         self.SetMenuBar(self._CreateMenu())
-        self.Bind(wx.EVT_MENU, self.OnOpenModel, id=self.ID_MODEL)
-        self.Bind(wx.EVT_MENU, self.OnNewShell, id=self.ID_SHELL)
-        self.Bind(wx.EVT_MENU, self.OnQuit, id=wx.ID_EXIT)
 
         # main panel
         self.panel = wx.Panel(self)
@@ -169,11 +116,16 @@ class Frame(wx.Frame):
         # files and/or a plugin system to add/enable/disable menu elements.
         menubar = wx.MenuBar()
         appmenu = wx.Menu()
-        appmenu.Append(self.ID_SHELL, '&New prompt\tCtrl+N', 'Open a new tab with a command prompt')
-        appmenu.Append(self.ID_MODEL, '&Open model\tCtrl+O', 'Open another model in a new tab')
+        menubar.Append(appmenu, '&App')
+        # Create menu items
+        shellitem = appmenu.Append(wx.ID_ANY,
+                                   '&New prompt\tCtrl+N',
+                                   'Open a new tab with a command prompt')
+        self.Bind(wx.EVT_MENU, self.OnNewShell, shellitem)
+        self.hook.menu(self, menubar)
         appmenu.AppendSeparator()
         appmenu.Append(wx.ID_EXIT, '&Quit', 'Quit application')
-        menubar.Append(appmenu, '&App')
+        self.Bind(wx.EVT_MENU, self.OnQuit, id=wx.ID_EXIT)
         return menubar
 
     def AddView(self, view, title):
@@ -191,9 +143,6 @@ class Frame(wx.Frame):
     def OnQuit(self, event):
         self.Close()
 
-    #----------------------------------------
-    # New command tab
-    #----------------------------------------
     def OnNewShell(self, event):
         self.NewCommandTab()
 
@@ -201,39 +150,3 @@ class Frame(wx.Frame):
         # TODO: create a toolbar for this tab as well
         # TODO: prevent this tab from being closed (?)
         self.notebook.AddPage(Crust(self.notebook), "Command", select=True)
-
-
-    #----------------------------------------
-    # Open new model:
-    #----------------------------------------
-    def load_model(self, mdata, **kwargs):
-        """Instanciate a new MadModel."""
-        res = mdata.repository.get()
-        model = MadModel(
-            name=mdata.name,
-            model=cpymad.model(mdata, **kwargs),
-            sequence=res.yaml('sequence.yml'))
-        return model
-
-    def show_model(self, madmodel):
-        view = MadLineView(madmodel)
-        panel = self.AddView(view, madmodel.name)
-        mirko = MirkoView(madmodel, view)
-
-        # create controller
-        MadCtrl(madmodel, panel, mirko)
-
-    def open_model(self):
-        from .openmodel import OpenModelDlg
-        dlg = OpenModelDlg(self)
-        success = dlg.ShowModal() == wx.ID_OK
-        if success:
-            model = dlg.data
-            h = os.path.join(self.logfolder, "%s.madx" % model.name)
-            self.show_model(self.load_model(model, histfile=h))
-        dlg.Destroy()
-        return success
-
-    def OnOpenModel(self, event):
-        self.open_model()
-
